can take driver from db and made driver and car repositories

+++ src/main/java/com/robosh/model/entities/Address.java
@@ -20,7 +20,7 @@ import javax.persistence.Id;
 public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer idAddress;
+    private Long idAddress;
 
     private String street;
     private String houseNumber;
